Find naive solution, and print nicer looking graphs & answer found.
Edges are curved, but need more control over location than the simple
multi-partite graph offers. This is probably OK for now, but I think
it's inspired me to manually place these nodes using a similar pattern.

My goal with the organization of the nodes is to see for myself the
patterns that emerge from n=4 to n=6. For n=4, there are 6 clusters of
4 nodes in the graph, with distances between clusters being > 1. When
finding a solution, the solver stays within a cluster, then hops to the
next nearest cluster. 4 out of the 5 jumps, it finds a cluster of
distance 2 away, but 1 out of the 5 jumps necessitates a jump of
distance 3, it appears. for N=5 my layout doesn't let me see it as
clearly and for n=6 it's impossible for me. Hence, I want to make this
better.

Next night of work..